+++ Gateway/Http/Client/ShoppingCartRefundClient.php
@@ -17,8 +17,10 @@ declare(strict_types=1);
 
 namespace MultiSafepay\ConnectCore\Gateway\Http\Client;
 
+use Exception;
 use Magento\Payment\Gateway\Http\ClientInterface;
 use Magento\Payment\Gateway\Http\TransferInterface;
+use Magento\Sales\Exception\CouldNotRefundException;
 use Magento\Store\Model\Store;
 use MultiSafepay\Api\Transactions\OrderRequest\Arguments\Description;
 use MultiSafepay\ConnectCore\Config\Config;
@@ -26,6 +28,7 @@ use MultiSafepay\ConnectCore\Factory\SdkFactory;
 use MultiSafepay\ConnectCore\Logger\Logger;
 use MultiSafepay\Exception\ApiException;
 use MultiSafepay\Exception\InvalidApiKeyException;
+use MultiSafepay\ValueObject\CartItem;
 use Psr\Http\Client\ClientExceptionInterface;
 
 class ShoppingCartRefundClient implements ClientInterface
@@ -73,6 +76,7 @@ class ShoppingCartRefundClient implements ClientInterface
     /**
      * @param TransferInterface $transferObject
      * @return array
+     * @throws Exception
      */
     public function placeRequest(TransferInterface $transferObject): array
     {
@@ -85,25 +89,35 @@ class ShoppingCartRefundClient implements ClientInterface
             $refundRequest = $transactionManager->createRefundRequest($transaction);
             $description = $this->description->addDescription($this->config->getRefundDescription($orderId));
             $refundRequest->addDescription($description);
-            $refundRequest->addMoney($request['money']);
 
             foreach ($request['payload'] as $refundItem) {
                 $refundRequest->getCheckoutData()->refundByMerchantItemId($refundItem['sku'], $refundItem['quantity']);
+            }
+
+            if (isset($request['adjustment'])) {
+                $refundRequest->getCheckoutData()->addItem(
+                    (new CartItem())->addName(__('Adjustment')->render())
+                    ->addDescription(__('Adjustment for order')->render())
+                    ->addQuantity(1)
+                    ->addUnitPrice($request['adjustment'])
+                    ->addMerchantItemId('adjustment_' . $request['credit_memo_id'])
+                    ->addTaxTableSelector('0')
+                );
             }
 
             return $transactionManager->refund($transaction, $refundRequest)->getResponseData();
         } catch (InvalidApiKeyException $invalidApiKeyException) {
             $this->logger->logInvalidApiKeyException($invalidApiKeyException);
 
-            return [];
+            throw new CouldNotRefundException(__($invalidApiKeyException->getMessage()));
         } catch (ApiException $apiException) {
             $this->logger->logExceptionForOrder($orderId, $apiException);
 
-            return [];
+            throw new CouldNotRefundException(__($apiException->getMessage()));
         } catch (ClientExceptionInterface $clientException) {
             $this->logger->logClientException($orderId, $clientException);
 
-            return [];
+            throw new CouldNotRefundException(__($clientException->getMessage()));
         }
     }
 }