+++ Test/Integration/Gateway/Request/ShoppingCartRefundRequestBuilderTest.php
@@ -99,6 +99,7 @@ class ShoppingCartRefundRequestBuilderTest extends AbstractTestCase
         $stateObject = new DataObject();
         $order = $this->getOrder();
         $order->getPayment()->setCreditMemo($this->getCreditMemo());
+        $order->getPayment()->getCreditMemo()->setAdjustment(10);
 
         $paymentDataObject = $this->getNewPaymentDataObject($order);
 
@@ -115,6 +116,8 @@ class ShoppingCartRefundRequestBuilderTest extends AbstractTestCase
         self::assertArrayHasKey('payload', $result);
         self::assertArrayHasKey('order_id', $result);
         self::assertArrayHasKey(Store::STORE_ID, $result);
+        self::assertArrayHasKey('credit_memo_id', $result);
+        self::assertArrayHasKey('adjustment', $result);
     }
 
     /**