+++ Test/Integration/Model/Api/Builder/OrderRequestBuilder/ShoppingCartBuilderTest.php
@@ -22,6 +22,7 @@ use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Sales\Model\Order;
 use MultiSafepay\Api\Transactions\OrderRequest;
+use MultiSafepay\Api\Transactions\OrderRequest\Arguments\CheckoutOptions;
 use MultiSafepay\ConnectCore\Model\Api\Builder\OrderRequestBuilder;
 use MultiSafepay\ConnectCore\Model\Api\Builder\OrderRequestBuilder\ShoppingCartBuilder;
 use MultiSafepay\ConnectCore\Model\Ui\Gateway\AfterpayConfigProvider;
@@ -58,7 +59,6 @@ class ShoppingCartBuilderTest extends AbstractTestCase
     {
         $orderRequest = $this->prepareOrderRequest();
         $orderRequest->addGatewayCode(AfterpayConfigProvider::CODE);
-
         self::assertArrayNotHasKey('shopping_cart', $orderRequest->getData());
     }